pyproject.toml
[project]
name = "llm-invoicing-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.109.1",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]
